+++ lib/equations.dart
@@ -9,7 +9,9 @@ export 'src/algebraic/types/laguerre.dart';
 export 'src/algebraic/types/linear.dart';
 export 'src/algebraic/types/quadratic.dart';
 export 'src/algebraic/types/quartic.dart';
+export 'src/algebraic/utils/algebraic_division.dart';
 export 'src/algebraic/utils/horner_results.dart';
+export 'src/algebraic/utils/polynomial_long_division.dart';
 export 'src/algebraic/utils/sylvester_matrix.dart';
 
 export 'src/nonlinear/nonlinear.dart';
@@ -46,5 +48,6 @@ export 'src/utils/exceptions/types/complex_exception.dart';
 export 'src/utils/exceptions/types/matrix_exception.dart';
 export 'src/utils/exceptions/types/nonlinear_exception.dart';
 export 'src/utils/exceptions/types/parser_exception.dart';
+export 'src/utils/exceptions/types/poly_long_division_exception.dart';
 export 'src/utils/exceptions/types/system_solver_exception.dart';
 export 'src/utils/expression_parser.dart';