+++ lib/src/nonlinear/types/chords.dart
@@ -31,7 +31,11 @@ class Chords extends NonLinear {
     required this.b,
     double tolerance = 1.0e-10,
     int maxSteps = 15,
-  }) : super(function: function, tolerance: tolerance, maxSteps: maxSteps);
+  }) : super(
+          function: function,
+          tolerance: tolerance,
+          maxSteps: maxSteps,
+        );
 
   @override
   bool operator ==(Object other) {