+++ lib/src/system/types/jacobi.dart
@@ -31,10 +31,11 @@ class JacobiSolver extends SystemSolver {
     this.maxSteps = 30,
     double precision = 1.0e-10,
   }) : super(
-            A: equations,
-            b: constants,
-            size: constants.length,
-            precision: precision) {
+          A: equations,
+          b: constants,
+          size: constants.length,
+          precision: precision,
+        ) {
     // The initial vector with the guesses MUST have the same size as the matrix
     // of course
     if (x0.length != size) {