+++ lib/src/utils/exceptions/types/parser_exception.dart
@@ -4,5 +4,8 @@ import 'package:equations/equations.dart';
 class ExpressionParserException extends EquationException {
   /// Represents an error for the [ExpressionParser] class.
   const ExpressionParserException(String message)
-      : super(message: message, messagePrefix: 'ExpressionParserException');
+      : super(
+          message: message,
+          messagePrefix: 'ExpressionParserException',
+        );
 }